Encapsulate user creation with a factory
