+++ src/Command/AddUserCommand.php
@@ -13,6 +13,7 @@ namespace App\Command;
 
 use App\Entity\User;
 use App\Repository\UserRepository;
+use App\User\UserFactory;
 use App\Utils\Validator;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Console\Command\Command;
@@ -22,7 +23,6 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
-use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 use Symfony\Component\Stopwatch\Stopwatch;
 
 /**
@@ -56,16 +56,16 @@ class AddUserCommand extends Command
     private $io;
 
     private $entityManager;
-    private $passwordEncoder;
+    private $userFactory;
     private $validator;
     private $users;
 
-    public function __construct(EntityManagerInterface $em, UserPasswordEncoderInterface $encoder, Validator $validator, UserRepository $users)
+    public function __construct(EntityManagerInterface $em, UserFactory $userFactory, Validator $validator, UserRepository $users)
     {
         parent::__construct();
 
         $this->entityManager = $em;
-        $this->passwordEncoder = $encoder;
+        $this->userFactory = $userFactory;
         $this->validator = $validator;
         $this->users = $users;
     }
@@ -181,16 +181,7 @@ class AddUserCommand extends Command
         // make sure to validate the user data is correct
         $this->validateUserData($username, $plainPassword, $email, $fullName);
 
-        // create the user and encode its password
-        $user = new User();
-        $user->setFullName($fullName);
-        $user->setUsername($username);
-        $user->setEmail($email);
-        $user->setRoles([$isAdmin ? 'ROLE_ADMIN' : 'ROLE_USER']);
-
-        // See https://symfony.com/doc/current/book/security.html#security-encoding-password
-        $encodedPassword = $this->passwordEncoder->encodePassword($user, $plainPassword);
-        $user->setPassword($encodedPassword);
+        $user = $this->userFactory->createUser($username, $email, $fullName, $plainPassword, [$isAdmin ? 'ROLE_ADMIN' : 'ROLE_USER']);
 
         $this->entityManager->persist($user);
         $this->entityManager->flush();
